refactor: change name to workshop
