--- src/pages/bootcamp/cursor/components/ui/utils.ts
@@ -1,6 +1,0 @@
-import {clsx, type ClassValue} from 'clsx'
-import {twMerge} from 'tailwind-merge'
-
-export function cn(...inputs: ClassValue[]) {
-  return twMerge(clsx(inputs))
-}
