+++ src/pages/workshop/ai/components/Features.tsx
