+++ src/pages/workshop/ai/components/Instructor.tsx
