+++ src/pages/workshop/ai/components/WorkshopStructure.tsx
