+++ src/pages/workshop/ai/components/animations.ts
