+++ src/pages/workshop/ai/components/ui/input.tsx
