+++ src/pages/workshop/ai/components/ui/utils.ts
