+++ src/pages/workshop/cursor/components/Features.tsx
