+++ src/pages/workshop/cursor/components/Instructor.tsx
