+++ src/pages/workshop/cursor/components/SignUpForm.tsx
@@ -31,18 +31,17 @@ export default function SignUpForm() {
 
     try {
       console.log('Submitting email:', email)
-
-      const currentDateTime = Math.floor(Date.now() * 0.001) // Customer.io uses seconds with their UNIX epoch timestamps
+      const currentDateTime = Math.floor(Date.now() * 0.001)
 
       if (!subscriberId) {
         identify.mutateAsync({
           email,
-          selectedInterests: {ai_bootcamp_waitlist: currentDateTime},
+          selectedInterests: {cursor_5day_workshop_waitlist: currentDateTime},
         })
       } else {
         identify.mutateAsync({
           id: subscriberId,
-          selectedInterests: {ai_bootcamp_waitlist: currentDateTime},
+          selectedInterests: {cursor_5day_workshop_waitlist: currentDateTime},
         })
       }
 
@@ -65,19 +64,20 @@ export default function SignUpForm() {
           className="max-w-2xl mx-auto"
         >
           <h2 className="mb-4 text-3xl font-bold text-center dark:text-white text-gray-900">
-            Ready to Build a Team of AI Devs?
+            Ready to Transform Your Cursor Workflow?
           </h2>
           <p className="mb-8 text-center text-gray-500 dark:text-gray-400 mx-auto">
-            Secure your spot in this unique, 20-day cohort-based workshop.
-            You&apos;ll learn alongside a supportive community of developers,
-            all on the same journey to master AI. Enter your email below and be
-            the first to know when registration opens.
+            Secure your seat in this 5-day, hands-on workshop designed to level
+            up your development process. Overcome the frustration of complex
+            integrations, learn to handle failures gracefully, and discover
+            powerful planning strategies to keep you shipping code with
+            confidence.
           </p>
           <form onSubmit={handleSubmit} className="max-w-md mx-auto">
             <div className="flex space-x-2">
               <Input
                 type="email"
-                placeholder="Enter your email to stay informed"
+                placeholder="Enter your email to join the waitlist"
                 value={email}
                 onChange={(e) => setEmail(e.target.value)}
                 required
@@ -87,15 +87,14 @@ export default function SignUpForm() {
               <Button
                 type="submit"
                 disabled={isSubmitting}
-                className="bg-blue-500  text-white  font-semibold transition-all duration-200 hover:scale-105 disabled:opacity-50 disabled:cursor-not-allowed"
+                className="bg-blue-500 text-white font-semibold transition-all duration-200 hover:scale-105 disabled:opacity-50 disabled:cursor-not-allowed"
               >
                 {isSubmitting ? 'Joining...' : 'Join Waitlist'}
               </Button>
             </div>
           </form>
           <p className="mt-4 text-center text-sm text-gray-500">
-            We&apos;ll send you all the details—no spam, just practical info on
-            how to join.
+            We&apos;ll only send you essential info about the course—no spam.
           </p>
         </motion.div>
       </div>
