+++ src/pages/workshop/cursor/components/WorkshopStructure.tsx
