+++ src/pages/workshop/cursor/components/animations.ts
