+++ src/pages/workshop/cursor/index.tsx
@@ -1,3 +1,4 @@
+'use client'
 import Hero from './components/Hero'
 import Features from './components/Features'
 import WorkshopStructure from './components/WorkshopStructure'
@@ -8,10 +9,10 @@ import type {NextPage} from 'next'
 import {cn} from '@/ui/utils'
 import {useTheme} from 'next-themes'
 
-const BootcampPage: NextPage = () => {
+const WorkshopPage: NextPage = () => {
   const {theme} = useTheme()
   console.log(
-    '[BootcampPage] Rendering on',
+    '[WorkshopPage] Rendering on',
     typeof window === 'undefined' ? 'server' : 'client',
   )
 
@@ -29,11 +30,13 @@ const BootcampPage: NextPage = () => {
 
         <div className="relative bg-[#111111]">
           <div
-            className={cn(
-              'absolute inset-0 bg-white dark:bg-gray-900',
-              theme === 'dark' && 'pattern-dots-dense',
-              theme === 'light' && 'pattern-dots-dense-light',
-            )}
+            className={`absolute inset-0 bg-white dark:bg-gray-900 ${
+              theme === undefined
+                ? ''
+                : theme === 'light'
+                ? 'pattern-dots-dense-light'
+                : 'pattern-dots-dense'
+            }`}
           />
           <Features />
         </div>
@@ -44,11 +47,13 @@ const BootcampPage: NextPage = () => {
 
         <div className="relative bg-[#111111]">
           <div
-            className={cn(
-              'absolute inset-0 bg-white dark:bg-gray-900',
-              theme === 'dark' && 'pattern-dots-dense',
-              theme === 'light' && 'pattern-dots-dense-light',
-            )}
+            className={`absolute inset-0 bg-white dark:bg-gray-900 ${
+              theme === undefined
+                ? ''
+                : theme === 'light'
+                ? 'pattern-dots-dense-light'
+                : 'pattern-dots-dense'
+            }`}
           />
           <Instructor />
         </div>
@@ -62,4 +67,4 @@ const BootcampPage: NextPage = () => {
   )
 }
 
-export default BootcampPage
+export default WorkshopPage
